COCHomeless: Datasets and Spatial Tools for U.S. Homelessness Research

Provides HUD Continuum of Care (CoC) Point-in-Time homeless counts, CoC geographic boundaries as simple-feature objects, a Census-to-CoC crosswalk linking census tracts and counties to CoCs, and county-level homeless estimates with American Community Survey covariates.

Data families

CoC counts

hud2007–hud2024 (and the preliminary hud2025_prelim).

CoC boundaries

coc2007–coc2024, national sf.

Census-to-CoC crosswalk

tract_coc2019, tract_coc_wt2019, county_coc2019 (and 2022).

County estimates

homeless, homeless_na, sp_homeless, area.

Citation

Cite the package with citation("COCHomeless"). Individual data types carry their own citations: the county-level estimates follow Almquist, Helwig and You (2020) doi:10.1080/08898480.2019.1636574 ; data-source credits for the Census-to-CoC crosswalk are given on the crosswalk help pages (tract_coc2019, county_coc2019).

See also

Author

Maintainer: Zack W. Almquist zalmquist@uw.edu (ORCID)

Other contributors:

Yun You [contributor]

Tom Byrne (Census-to-CoC crosswalk procedure) [contributor]